Add graphql layer and refactor directory structure
